Raise error if uploaded album is not a zip or tar
Archives other than .zip and .tar.gz are rejected, even though the
tarfile library supports files with lzma and bzip2 compression, as the
MIME types for these compressions are not registered and the values in
request.content_type are user supplied, so these may vary. Perhaps we
should use libmagic in the future to avoid these issues.