Really fix the statistics this time
Members with `Membership`s of multiple types (e.g. one `Membership` as a normal
member and one `Membership` as a honorary member) were counted twice in the
statistics due to a `membership__type=...` in the queries.

The fix is counting the number of active `Membership`s instead of the number of
active `Member`s (in a given study year).