Fix sorting by payments in events admin
Rewrote the payment sorting parser and changed the sort behaviour when
changing payments. Now the table will not reorder when changing a
payment, as this happens after a network call and can cause an
unexpected moving of the edited row. Now instead the table will not be
resorted, and instead is reverted to an unordered state.