Remove pk and add slug to serializers
Replaces the pk with the slug in the album serializers.
Removes the album pk from the Photo serializer, since there is no
way to access an album from pk any more. We also don't need to have
the pk, since there is no endpoint that exposes only a Photo, without
the album (with slug) that it is part of.