Cache Member.latest_membership
This solves N+1 queries where a single users's latest_membership is used
many times. For example, in /api/v2/events/, where it's used once for
each event, to get the permissions for that event.

This caching can be risky, as it's never reset. But the same goes for
any _latest_membership annotation from #3229.

I still need to ensure that this doesn't break the registrations app.