Initial commit
This commit adds and initializes to some files. We add courses.js, which includes some jQuery which filters through all of the rows of the table by the input of the search bar. We check which class-name has a part in it close to the search term, then go to the closest parent, and remove these from the DOM.