Fix renewal moneybird invoice period for upgrages (#3956)
Membership upgrades got pushed with the membership.since date as period, which can be a year ago, in a locked period.
This changes it to the payment date, which is normally the start of when the renewal/registration takes effect.

For upgrades, this is the moment the upgrade happens.
For renewals, this is still also the start of the new membership.
For registrations, it's also the start of the membership.